Move `EventLogger` to separate file
