+++ Tests_v3/EventLogger.swift
@@ -1,0 +1,21 @@
+//
+// Copyright © 2020 Stream.io Inc. All rights reserved.
+//
+
+import Foundation
+@testable import StreamChatClient
+
+/// The type can be used to check the events published by `NotificationCenter`
+final class EventLogger {
+    @Atomic var events: [Event] = []
+    var equatableEvents: [EquatableEvent] { events.map(EquatableEvent.init) }
+    
+    init(_ notificationCenter: NotificationCenter) {
+        notificationCenter.addObserver(self, selector: #selector(handleNewEvent), name: .NewEventReceived, object: nil)
+    }
+    
+    @objc
+    func handleNewEvent(_ notification: Notification) {
+        events.append(notification.event!)
+    }
+}
